Pass dmsetup table via stdin
Previously, we were passing keys as command-line arguments to
dmsetup.  Obviously, this is bad, because a simple 'ps' command
would leak the key.

Instead, pass the keys via stdin to dmsetup.

(This quite obvious issue was on my list prior to version 1.0.0,
but it seems I was too eager to get the first release done and
forgot to address it.)